requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
ariadne==0.23.0
cryptography==43.0.3
